Updated ui to properly update articles into zinc
