+++ src/components/articles-view/index.jsx
@@ -3,9 +3,30 @@ import { useParams } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
 import "./article.css"
 import Article from "./components/article";
+import Box from '@mui/material/Box';
+import Modal from '@mui/material/Modal';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
+import { useState } from "react";
+const style = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 400,
+    bgcolor: 'background.paper',
+    border: '2px solid #000',
+    boxShadow: 24,
+    p: 4,
+    color: "black"
+  };
 const ArticlesView = ({}) => {
     const { uuid } = useParams();
     const navigate = useNavigate();
+    const [open, setOpen] = useState(false);
+    const handleOpen = () => setOpen(true);
+    const handleClose = () => setOpen(false);
+    const role = sessionStorage.getItem("role");
     const articlesQuery = gql`
     query articles($articlesInput: ArticlesPrivate){
         articlesPrivate(input: $articlesInput){
@@ -22,7 +43,8 @@ const ArticlesView = ({}) => {
                 tags{
                     tag
                 }
-                }
+            }
+            zincPassword
         }
     }
     `
@@ -54,11 +76,36 @@ const ArticlesView = ({}) => {
     if (loading) return <p>Loading Graphql data...</p>
     
     if (error) return `Submission error! ${error.message}`;
+    console.log(data)
     return(
         <>
         <button onClick={() => {
             navigate(`/projects/${uuid}/article-creation`)
         }}>Create Article</button>
+        {role == "Admin" ?
+        <> 
+        <button onClick={handleOpen}>Show Password</button>
+            <Modal
+                open={open}
+                onClose={handleClose}
+                aria-labelledby="modal-modal-title"
+                aria-describedby="modal-modal-description"
+            >
+                <Box sx={style}>
+                <Typography id="modal-modal-title" variant="h6" component="h2">
+                    Credentials for searching
+                </Typography>
+                <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+                    {`Username:${uuid}`}
+                </Typography>
+                <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+                    {`Password:${data.articlesPrivate.zincPassword}`}
+                </Typography>
+                <Button onClick={() => handleClose()}>Close</Button>
+                </Box>
+            </Modal>
+            </> : <></> 
+            }
         {data.articlesPrivate.article.length > 0 ?< button onClick={() => {
             deleAllArticles({
                 variables:{
@@ -69,7 +116,6 @@ const ArticlesView = ({}) => {
         <button onClick={() => {
             navigate(-1);
         }}>Go Back</button>
-
         {data.articlesPrivate.article.map((article) => (
             <div className="article" key={article.uuid}>
             <Article articleUuid={article.uuid} articleName={article.title} key={article.uuid} articleDescription={article.description} author={article.author.name} ArticleData={article}/>
